hook up GB callbacks to frontend
hooked all callbacks, only vblank and error is fully impl.

vblank callback is to let the frontend know that it should render the screen. this removes any screen tearing and changes to the pixels whilst in vblank.

error callback is to be used to handle errors that may happen during emulation. these will be updated with warnings also, such as trying to write to a rom or OOB read / write from the game.