fixed dual windows not closing
so when theres multiple windows (in sdl2) open, sdl quit will no longer trigger once a window is selected to be closed...which makes sense.

instead a window event is sent, with type close. note however that this event is sent even when theres only 1 window, so this event will trigger along with sdl_quit if its the last window