added mbc3 rtc support
fixed bugs in my mbc table array, i had mbc5 set as rtc, when only mbc3 supports it.

theres a speedhack for rtc that ive added, basically, all pokemon games and harvest moon read from the same addr when accessing rtc, 0xA000. because of this, a pointer to that reg can be stored at that addr.

this is very hacky, and if the addr goes out of range, it'll break.

this is totally fine for the games i have listed, and theres not that many rtc games so it's reasonable to test them all, though, this will break any homebrew games that read from another addr