start work on GB printer!
printer support is good enough that it manages to successfully *print* any pokemon from the pokemon crystal pokedex.

however, i have not checked the pixel data output yet, and storing the pixels to ram is currently broken.

atm, printer checksum fails for when big data is sent.

also, the serial interface thats used for pokemon 2-player trading does not work correctly, so it had to be changed. i need to re-write the interface so that it works for both devices...