add apu struct, start to replace IO array with struct bitfields
decided the IO array wasn't worth it because of the apu regs needed to be manipulated a lot based on bits, and i'd rather not litter the code with shifts and masks everywhere. instead, ive placed it all in one place, read / write bus.

this might impact performance, but likely it's very minor.

will start to remove io array slowly overtime, but the helper macros (IO) will stay.

Fixes #7