add all apu channels and effects!
still a work-in-progress.

the sound for channel 1 and 2 sounds pretty good! channel 4 (noise) is a hack using rand, but it sounds good enough for now. channel 3 (wave) sounds awful, very very bad.

i am doing no volume mixing, so the wave channel can get really loud, much louder than everything else, for this, i have removed the 100% volume option from it so that it sounds better for now.

this is just a POC, just something that sounds acceptable for now.