improved audio, removed lz4
made lots of changes to audio.

fixed the noise channel freq. i read the doc very wrong, i have no idea how i came to the first conclusion, mustve been a late night...

using negative values for if the duty cycle is 0.

removed testing audio callback, added a slightly better one which drives the emu core from the audio thread. this works fine but has a very race conditions with the current "frontend", really though the frontend was not designed with this in mind, so it'll be worked on later on.

fixed a few wrong return values for the apu io. some values were write only, so now they return 0xFF, and a mask was wrong in one case.

i removed the padding from the structs for the apu, was no need for it.

lz4 was an alternitive to zstd for compressing savestates and rewind buffers, however zstd was always faster and compressed better, so im removing it.