use normal stdint and bool types instead of my own typedef types
this replaces all GB_U8 with uint8_t

this also found a bug with `GB_get_sprite_size` as it returned a bool rather than a u8. however the GB_BOOL was actually just a u8, so the compiler didnt warn. however now it warned and i changed the return type to u8.