added basic emscripten port
so far its fully functional it terms of it can load roms and plays sound. however theres no saving yet.

i plan to use cloud storage for saves and savestates, so it can be easily shared across browsers. also keep a copy of the savefile in browser cache just in case the save fails to upload to the cloud or loses internet connection at anypoint (or abruptly closes the browser)