fix gbc bg-oam priority, use correct initial cpu reg values
issue with gbc bg was that i read the docs wrong and thought it was saying that the bg prio only applies to tiles in bank-0 of vram. however that was a mistake.

fixed a comment about mbc2 being 512kb, its actually 512 bytes, the code around the comment was correct however.

made a note about needing to add halt bug behavior at some point