add build-time option for setting any pixel format
this allows for setting 8,16,32 bit pixels, and allowing the user to
set a ptr into the pixel buffer themselves.

this will be a nice speedup for software rendering, as pixels can be
written directly to the screen

furthermore, this is useful for systems with strange display setups
such as one embeded system i worked on used 256 LUT for screen pixels
which actually just indexed into a colour array, so the LUT acted as
colour palette table.

this is not yet finished, gbc support needs to be added by either
adding a callback on palette swap, so the user can mix the colours
themelves, or the core can handle it.

i think the callback is better suited as it's really out of scope for a
gb core to be in charge of colour mixing...