started work on zrom support
this is a basic rom compression idea i had, ideal for systems with tiny amounts
of storage (ie, embeded) in which the roms will be first compressed using
my zrom cli tool (not commited yet) which compresses each rom bank (16KiB)
with the header and entries field at the top of the file containing info
on each compressed bank.

i chose lz4 as the compressor as it's simple, fast, and uses no memory
allocation!
I also tried lz4hc, however some roms would actually compress *worse*
then compaired to lz4, for example all pokemon games compressed worse.

some banks will not compress well, and even end up bigger than 16KiB!
for this, we do not compress that bank, and instead store it raw, and
set the flag to UNCOMPRESSED.
this means when on bank swap, a simple memcpy is all that's needed.

a potential option may be to have the user set the mempool, which could
contain > 2 banks, which would allow for zrom to uncompress that many
banks upfront, potentially saving many lz4_decompress calls!

i have not tested this for speed yet, so see if its worth doing this...