remove mbc write and get_rom/bank func ptrs, replace with enum switch
reason for the change is because the old mbc setup table took 8*4*256 of space,
which is 8kib, when really, only a few places were taken.

so i decided to just have 2 vars in the table, 1 for flags, 1 for the mbc type.
the idea was to have mbcX_init() funcs which setup the func pointers, but
i realised that using func pointers here serve no purpose. instead, using a
switch enum for swapping banks and ram writes, all of which can now be inlined!