allow user to set sram, rather than having sram being part of gb struct
this means that the way a rom should be loaded is like this:

1. `GB_get_rom_info(rom_data)`
  this gets the info such as ram_size and mbc flags.
  the user can then see if the game uses sram, and alloc the needed ram for
  it, and set it via `GB_set_sram(sram)`.
  the user can also check if the game has battery saved ram, if so, they can
  load the save into a buffer and set it with the func above.

  finally, this allows for the option to use mmap saves now!

2. `GB_loadrom(rom_data)`
  this loads the rom, performs a few checks (such as has enough sram)
  and then sets up the gb struct based on the system the game requires.

this is a bit more manual than most other emus, as they often have a simple
`loadrom(file_path)` function, which uses stdio, malloc and such.

while thats great, it locks you into using the emu in a certain way and becomes
more difficuilt to port to baremetal embedded systems without having to patch
the emu in multiple places.