add zrom support, enable building with/out GBC and SGB support
zrom support was added to support embedded systems that have very small
amount of flash mem (1MiB or less) and so cannot fit roms, such as pokemon red.

zrom basically keeps bank0 uncompressed, and as the first 16KiB in the file.
this allows for all of the header funcs to still work.

after the 16KiB, then the zrom header, then the zrom entries (for each bank).

building without GBC (and SGB but that wasn;t finished anyway) support is now
a build time option.

this again, was a change for embedded systems that struggle with low ram, but
have enough ram for DMG, but not the extra ~32KiB that GBC needs for wram
and vram.

the zrom cli code isn't uploaded yet, but will do in next few commits