optimisations!
for ppu rendering, the render functions were turned into (poor mans template) macros
to support different bpp.
this removes the overhead of a switch on each pixel write.

when sprite fetching, only fetch sprite_y first, and if in range, then fetch the
rest of the data from oam.

break early from render loop if the tile is offscreen.

for ALU REG_A instructions, use a simpler instruction for each, for example,
OR_A_A will not change A, only the flags will be set. SUB_A_A, A = 0 etc...

as the compiler was inlining basically everything inside the run_step() function
(making it huge), it therefore not inline that step function if it was called in more
than one place, which is was.
because of this, there was a function call on each step.
for now, i have just removed the other instance where step was called.
this was a nice 100+fps boost.

set apu regs value on reset