fix apu sweep, len, envelope bugs
i was incorrectly disabling envelope if period=0 when clocking envelope.
this was only noticiable in prehistorik man, where sq 1/2 where disabled
and only noise channel played.

when and how sweep gets clocked has changed.
previously, on trigger, sweep calc was only performed if sweep shift was non zero
and sweep was enabled, but this isn't the case. it's always performed, as long as
sweep shift is non zero.

a bug in nr34 where extra len clock would happen caused the len counter to be decremented twice!
this was probably a copy paste bug as theres 3 other len functions that did exactly what is was
meant to...

dmg_sound: fail 9, 10, 12
cgb_sound: fail 9, 12

[skip-ci]